Resolve show all annotations bug
Commit resolves bug that does not show all annotations when switching for an individual property to all annotations whilst having the side panel open.
